requirements.txt
altgraph==0.17.4
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
idna==3.10
macholib==1.16.3
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pyzbar==0.1.9
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0